% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common_UniformSav.R
\name{addLogMessages}
\alias{addLogMessages}
\title{Add Log Message}
\usage{
addLogMessages(...)
}
\arguments{
\item{...}{character string of the message, use curly brackets (braces) to input objects (glue notation)}
}
\description{
Add a vector of log messages to the global log message queue
}
