% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_Main.R
\name{annualOutlier}
\alias{annualOutlier}
\title{function to check if annual limits are exceeded by angler over the month.
parameters are:
catch = responses where Chinook, Coho, Lingcod and Halibut were encountered
ANN_CN = Annual limit for Chinook based on the UI values provided
ANN_LC = Annual limit for Lingcod based on the UI values provided
ANN_HA = Annual limit for Halibut based on the UI values provided}
\usage{
annualOutlier(catch, ANN_CN, ANN_LC, ANN_HA)
}
\description{
function to check if annual limits are exceeded by angler over the month.
parameters are:
catch = responses where Chinook, Coho, Lingcod and Halibut were encountered
ANN_CN = Annual limit for Chinook based on the UI values provided
ANN_LC = Annual limit for Lingcod based on the UI values provided
ANN_HA = Annual limit for Halibut based on the UI values provided
}
