% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_Main.R
\name{areaOutlier}
\alias{areaOutlier}
\title{function to check for multiple areas fished and their likelihood based on a predefined "neighbours" table
parameters are:
SavDat = Dataframe of response data, processed for format and adjusts/excludes
areas_df = dataframe for to map iREC reponse areas to REF data areas using CREST mappings
nn = dataframe of acceptable area combinations for same day fishing.
areaout_returns = boolena flag to return all repeating areas fished patters or not}
\usage{
areaOutlier(SavDat, areas_df, nn, areaout_returns)
}
\description{
function to check for multiple areas fished and their likelihood based on a predefined "neighbours" table
parameters are:
SavDat = Dataframe of response data, processed for format and adjusts/excludes
areas_df = dataframe for to map iREC reponse areas to REF data areas using CREST mappings
nn = dataframe of acceptable area combinations for same day fishing.
areaout_returns = boolena flag to return all repeating areas fished patters or not
}
