% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KREST_common.R
\name{Authenticate}
\alias{Authenticate}
\title{wrapper function to authenticate with the KREST API}
\usage{
Authenticate(username, password)
}
\description{
wrapper function to authenticate with the KREST API
}
