% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_Main.R
\name{BuildCatch}
\alias{BuildCatch}
\title{Function to pull Chinook, Coho, Halibut and lingcod catches out of the normalized response data
Parameters are:
responses = normalized response data tid, surveykey, item, val
resp = the number of anglers (juveniles), areas and dates reported fished}
\usage{
BuildCatch(responses, resp)
}
\description{
Function to pull Chinook, Coho, Halibut and lingcod catches out of the normalized response data
Parameters are:
responses = normalized response data tid, surveykey, item, val
resp = the number of anglers (juveniles), areas and dates reported fished
}
