% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates_KREST.R
\name{buildJSONS}
\alias{buildJSONS}
\title{The data sent to KREST is sent as a JSON string.
This funcion transforms the dataframe to a JSON string
parameters are:
rslt = the dataframe containing the formatted for KREST estimates
inds = the components of the file name in the form Month_Year_... like "May_2026_Result_20260723_092514.xlsx"
out_path_in = Optional-- the name of the path where we would write an xlsx form of the data that will form the individual
data packets sent to KREST.}
\usage{
buildJSONS(rslt, inds, out_path_in = "")
}
\description{
The data sent to KREST is sent as a JSON string.
This funcion transforms the dataframe to a JSON string
parameters are:
rslt = the dataframe containing the formatted for KREST estimates
inds = the components of the file name in the form Month_Year_... like "May_2026_Result_20260723_092514.xlsx"
out_path_in = Optional-- the name of the path where we would write an xlsx form of the data that will form the individual
data packets sent to KREST.
}
