% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_Main.R
\name{buildOutlierXLS}
\alias{buildOutlierXLS}
\title{function to build an XLSX file from the output of the function processSav and crazy8s
parameters are:
n_dfs = list of dataframes returned from processSav
flagged_df = dataframe returned from crazy8's code}
\usage{
buildOutlierXLS(n_dfs, flagged_df)
}
\description{
function to build an XLSX file from the output of the function processSav and crazy8s
parameters are:
n_dfs = list of dataframes returned from processSav
flagged_df = dataframe returned from crazy8's code
}
