% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_Main.R
\name{BuildProfanityLists}
\alias{BuildProfanityLists}
\title{Function to build list of all the nasty works we can find.
No parameters needed}
\usage{
BuildProfanityLists()
}
\description{
Function to build list of all the nasty works we can find.
No parameters needed
}
