% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_Code.R
\name{buildsql}
\alias{buildsql}
\title{function to adjust a SQL where statement based on if there is already a where clause in the statement.
Parameters are:
sqlin = base sql statment
field = name of the field to add a where condition to
vals = the value(s) to select on}
\usage{
buildsql(sqlin, field, vals)
}
\description{
function to adjust a SQL where statement based on if there is already a where clause in the statement.
Parameters are:
sqlin = base sql statment
field = name of the field to add a where condition to
vals = the value(s) to select on
}
