% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_Code.R
\name{can_be_integer}
\alias{can_be_integer}
\title{function to check if a character(string?) can be an integer value
parameters are:
chr = string to test if it can represent an integer/number}
\usage{
can_be_integer(chr)
}
\description{
function to check if a character(string?) can be an integer value
parameters are:
chr = string to test if it can represent an integer/number
}
