% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_Main.R
\name{catchOutlier}
\alias{catchOutlier}
\title{function to check for reported kept catch in excess of daily bag limits as well as reported large total encounters (kept + released)
parameters are:
catch = responses where Chinook, Coho, Lingcod and Halibut were encountered
areas_df = dataframe for to map iREC responses areas to REF data areas using CREST mappings
HABAG = Halibut daily bag limit based on the UI values provided
CNBAG = Chinook daily bag limit based on the UI values provided
COBAG = Coho daily bag limit based on the UI values provided
ENCBAG = Kept + Released daily limit based on UI values provided used for all 4 species
fris_data = dataframe of bag limits by month, area and species - used in place of UI limits
FRIS_GRACE = integer number of fish in excess of FRIS limits to allow.}
\usage{
catchOutlier(
  catch,
  areas_df,
  HABAG,
  CNBAG,
  COBAG,
  ENCBAG,
  fris_data,
  FRIS_GRACE
)
}
\description{
function to check for reported kept catch in excess of daily bag limits as well as reported large total encounters (kept + released)
parameters are:
catch = responses where Chinook, Coho, Lingcod and Halibut were encountered
areas_df = dataframe for to map iREC responses areas to REF data areas using CREST mappings
HABAG = Halibut daily bag limit based on the UI values provided
CNBAG = Chinook daily bag limit based on the UI values provided
COBAG = Coho daily bag limit based on the UI values provided
ENCBAG = Kept + Released daily limit based on UI values provided used for all 4 species
fris_data = dataframe of bag limits by month, area and species - used in place of UI limits
FRIS_GRACE = integer number of fish in excess of FRIS limits to allow.
}
