% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates_KREST.R
\name{checkEstExists}
\alias{checkEstExists}
\title{function to test if a named estimate exists
parameters are:
estname = the named estimate we are looking for
krestroot = the database url we are interested in.
returns either false if no estimate name matches
else returns the internal KREST id for the first estimate matching that name.}
\usage{
checkEstExists(estname, krestroot)
}
\description{
function to test if a named estimate exists
parameters are:
estname = the named estimate we are looking for
krestroot = the database url we are interested in.
returns either false if no estimate name matches
else returns the internal KREST id for the first estimate matching that name.
}
