% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_Main.R
\name{CheckProfanity}
\alias{CheckProfanity}
\title{Function to build the parameters and then check the comments for sentiment and profanity
Parameters are:
SavDat = reponse data}
\usage{
CheckProfanity(SavDat)
}
\description{
Function to build the parameters and then check the comments for sentiment and profanity
Parameters are:
SavDat = reponse data
}
