% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates_KREST.R
\name{deleteEst}
\alias{deleteEst}
\title{function to delete an estimate based on its internal KREST ID
parameters are:
est_id = the internal KREST estimate ID
krestroot = the database url we are interested in.}
\usage{
deleteEst(est_id, krestroot)
}
\description{
function to delete an estimate based on its internal KREST ID
parameters are:
est_id = the internal KREST estimate ID
krestroot = the database url we are interested in.
}
