% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_Code.R
\name{Expmonth}
\alias{Expmonth}
\title{Function to return pretty version of month
parameters are:
mIn = month to convert.  Can be a number or short or long version of the month name}
\usage{
Expmonth(mIn)
}
\description{
Function to return pretty version of month
parameters are:
mIn = month to convert.  Can be a number or short or long version of the month name
}
