% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates_KREST.R
\name{FetchGrid}
\alias{FetchGrid}
\title{function to fetch the data grid view of the available estimates from the KREST UI server
the filter set is based on what was previously sent by the setFilter function.
parameters are:
krestroot = the database url we are interested in.}
\usage{
FetchGrid(krestroot)
}
\description{
function to fetch the data grid view of the available estimates from the KREST UI server
the filter set is based on what was previously sent by the setFilter function.
parameters are:
krestroot = the database url we are interested in.
}
