% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses_CREST.R
\name{filterRespColumns}
\alias{filterRespColumns}
\title{function to filter response dataset based on list of column names}
\usage{
filterRespColumns(survey_data, cols)
}
\description{
function to filter response dataset based on list of column names
}
