% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_Code.R
\name{fixAdjDates}
\alias{fixAdjDates}
\title{function to check and fix if necessary the date formats in an adjust csv file
paramters are:
adj_df = the adjust csv file loaded into a dataframe
CSVfilename = the name of the adjust file}
\usage{
fixAdjDates(adj_df, CSVfilename)
}
\description{
function to check and fix if necessary the date formats in an adjust csv file
paramters are:
adj_df = the adjust csv file loaded into a dataframe
CSVfilename = the name of the adjust file
}
