% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_Code.R
\name{fixComments}
\alias{fixComments}
\title{fix the comments parsing for email adress and phone numbers and emoji's}
\usage{
fixComments(survey_data, cmnt = "COMMENT")
}
\description{
fix the comments parsing for email adress and phone numbers and emoji's
}
