% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_Main.R, R/common_Code.R
\name{fixDates}
\alias{fixDates}
\title{function to see if the date is a date, or if it is an integer representing the number of days since
January 1, 1970 - I think this is from old format data
parameters are:
SavDat = response file with item = "datefished" as the date field to check/correct}
\usage{
fixDates(SavDat)

fixDates(SavDat)
}
\description{
function to see if the date is a date, or if it is an integer representing the number of days since
January 1, 1970 - I think this is from old format data
parameters are:
SavDat = response file with item = "datefished" as the date field to check/correct

function to see if the date is a date, or if it is an integer representing the number of days since
January 1, 1970 - I think this is from old format data
parameters are:
SavDat = response file with item = "datefished" as the date field to check/correct
}
