% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_Code.R
\name{fixValues}
\alias{fixValues}
\title{fix common differences in what is delivered vs what is expected}
\usage{
fixValues(survey_data)
}
\description{
fix common differences in what is delivered vs what is expected
}
