% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates_CREST.R
\name{formatdata}
\alias{formatdata}
\title{Function to translate the XLSX output from the iREC analysis into
normalized data tables in CREST
parameters are:
dfs = list of dataframes holding the detailed estimates, detailed variances and summarized variances
SHEET_NAMES = list of the XLSX file tab names for the data
inds = a list of the comonents for the xlsx filename split on the "-" character to determin the month and year the estimates
represent
src_id = the primary key of the creel_irec_source table where the xlsx file is stored
dbCon = an oracle connection to the database destination}
\usage{
formatdata(dfs, SHEET_NAMES, inds, src_id, dbCon)
}
\description{
Function to translate the XLSX output from the iREC analysis into
normalized data tables in CREST
parameters are:
dfs = list of dataframes holding the detailed estimates, detailed variances and summarized variances
SHEET_NAMES = list of the XLSX file tab names for the data
inds = a list of the comonents for the xlsx filename split on the "-" character to determin the month and year the estimates
represent
src_id = the primary key of the creel_irec_source table where the xlsx file is stored
dbCon = an oracle connection to the database destination
}
