% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_Code.R
\name{get_mode}
\alias{get_mode}
\title{Function to calculate the mode from a list
parameters are:
x = list to calculate the modal value from}
\usage{
get_mode(x)
}
\description{
Function to calculate the mode from a list
parameters are:
x = list to calculate the modal value from
}
