% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates_KREST.R
\name{getandformatKData}
\alias{getandformatKData}
\title{function to load and fully format the iREC estimates - Calls loadXLSX to actually load the data and do some of the standardizing
but this does the heavy lifting in terms of formatting
parameters are:
inds = the components of the file name in the form Month_Year_... like "May_2026_Result_20260723_092514.xlsx"
xls_file = the name of the excel sheet to use
xls_path = a list of paths to search for the file in
SHEET_NAMES = the named tabs that holds the data wee need
source_id = the primary key of the record in the table OTOLITH_V1.CREEL_IREC_SOURCE to fetch the source excel file from
returns a list of data products:
rslt position one <- TRUE - presumably a flag that everything worked
rslt position two <- t_est_g - detailed estimates with variance formatted for KREST
rslt position three <- xls_file - the name of the file the estimates came from
rslt position four <- csvs - the look-up tables used
rslt position five <- St_est_g - The summary estimates with variance formatted for KREST}
\usage{
getandformatKData(inds, xls_file, xls_path, SHEET_NAMES, source_id)
}
\description{
function to load and fully format the iREC estimates - Calls loadXLSX to actually load the data and do some of the standardizing
but this does the heavy lifting in terms of formatting
parameters are:
inds = the components of the file name in the form Month_Year_... like "May_2026_Result_20260723_092514.xlsx"
xls_file = the name of the excel sheet to use
xls_path = a list of paths to search for the file in
SHEET_NAMES = the named tabs that holds the data wee need
source_id = the primary key of the record in the table OTOLITH_V1.CREEL_IREC_SOURCE to fetch the source excel file from
returns a list of data products:
rslt position one <- TRUE - presumably a flag that everything worked
rslt position two <- t_est_g - detailed estimates with variance formatted for KREST
rslt position three <- xls_file - the name of the file the estimates came from
rslt position four <- csvs - the look-up tables used
rslt position five <- St_est_g - The summary estimates with variance formatted for KREST
}
