% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KREST_common.R
\name{getAuthToken}
\alias{getAuthToken}
\title{function to build and send the web call to the KREST server identified in the variable "krestendpoint"
to get the authorization token - Token lives for about 10 mins
parameters are:
krestendpoint = KREST server test vs prod
Logincreds = structure passed to KREST contains username and password of KREST account to use
logincreds <<- list(grant_type = "password",
password = "password",
username = "username")
returns the bearer token for use in transactions}
\usage{
getAuthToken(krestendpoint, logincreds)
}
\description{
function to build and send the web call to the KREST server identified in the variable "krestendpoint"
to get the authorization token - Token lives for about 10 mins
parameters are:
krestendpoint = KREST server test vs prod
Logincreds = structure passed to KREST contains username and password of KREST account to use
logincreds <<- list(grant_type = "password",
password = "password",
username = "username")
returns the bearer token for use in transactions
}
