% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common_UniformSav.R
\name{getCatchColNames}
\alias{getCatchColNames}
\title{Get Catch Column Names}
\usage{
getCatchColNames(survey_df)
}
\arguments{
\item{survey_df}{A survey related data frame}
}
\value{
A vector of column names related to catch, including effort_days and juv_effort_days
}
\description{
Retrieve a vector of catch column names from the survey data frame provided
}
