% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_Main.R
\name{GetCompoundProfanity}
\alias{GetCompoundProfanity}
\title{Function to build a compound word profanity dictionary
seems not so nice to leave a list of nasty words for someone to stumble upon
No parameters needed}
\usage{
GetCompoundProfanity()
}
\description{
Function to build a compound word profanity dictionary
seems not so nice to leave a list of nasty words for someone to stumble upon
No parameters needed
}
