% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_Code.R
\name{getdata}
\alias{getdata}
\title{Function to retrieve data from a database connection (Oracle)
parameters are:
Kcon = database connection returned from "setupOracleConn"
qry = sql statement for the data required
chunks  = boolean to retrieve the data in blocks of 25 rows at a time.
Once all the "chunks" are retieve the merged dataframe is returned.
Total time to retrieve the data is displayed in the console}
\usage{
getdata(Kcon, qry, chunks = FALSE)
}
\description{
Function to retrieve data from a database connection (Oracle)
parameters are:
Kcon = database connection returned from "setupOracleConn"
qry = sql statement for the data required
chunks  = boolean to retrieve the data in blocks of 25 rows at a time.
Once all the "chunks" are retieve the merged dataframe is returned.
Total time to retrieve the data is displayed in the console
}
