% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KREST_common.R
\name{GetLookupCodes}
\alias{GetLookupCodes}
\title{function to fetch look-up tables from KREST}
\usage{
GetLookupCodes(
  BearerToken,
  resource,
  dateTime = NULL,
  showAssignmentHistory = FALSE
)
}
\description{
function to fetch look-up tables from KREST
}
