% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KREST_common.R
\name{GetOAuth2Token}
\alias{GetOAuth2Token}
\title{function to get authorization token from KREST api}
\usage{
GetOAuth2Token(username, password)
}
\description{
function to get authorization token from KREST api
}
