% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_Main.R
\name{GetProfanityScores}
\alias{GetProfanityScores}
\title{Function to get the profanity scores and merge in the sentiment scores
parameters are:
SavDat = response data
lexicons = nasty word list
nrc_data = previously calculated emotions and valence from NRC Dictionary
sentiment_scores = previously calculated sentiment score for each comment}
\usage{
GetProfanityScores(SavDat, lexicons, nrc_data, sentiment_scores)
}
\description{
Function to get the profanity scores and merge in the sentiment scores
parameters are:
SavDat = response data
lexicons = nasty word list
nrc_data = previously calculated emotions and valence from NRC Dictionary
sentiment_scores = previously calculated sentiment score for each comment
}
