% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_Main.R
\name{GetSentiment}
\alias{GetSentiment}
\title{Function to assess the comments for sentement - How angry or happy the comments seem to be
Uses the package syuzhet
parameters are:
SavDat = raw response data}
\usage{
GetSentiment(SavDat)
}
\description{
Function to assess the comments for sentement - How angry or happy the comments seem to be
Uses the package syuzhet
parameters are:
SavDat = raw response data
}
