% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_Code.R
\name{GetYearMonth}
\alias{GetYearMonth}
\title{Function to determine the year an month of a response file including adjusted responses
adjust file can introduce date records with different formats
parameters are:
dfIn = response file with fields did_not_fish, datefished}
\usage{
GetYearMonth(dfIn)
}
\description{
Function to determine the year an month of a response file including adjusted responses
adjust file can introduce date records with different formats
parameters are:
dfIn = response file with fields did_not_fish, datefished
}
