% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common_UniformSav.R
\name{labelText}
\alias{labelText}
\title{Columns read from an SPSS file with the haven R package provides labelled columns.
This function converts the labels (that are kind of like factors) into a text column.}
\usage{
labelText(values)
}
\arguments{
\item{values}{A vector of labelled values}
}
\value{
Text vector
}
\description{
Columns read from an SPSS file with the haven R package provides labelled columns.
This function converts the labels (that are kind of like factors) into a text column.
}
