% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses_CREST.R
\name{link_Lus}
\alias{link_Lus}
\title{function to link look-ups to the SAV file in dataframe form}
\usage{
link_Lus(df_sav, lus)
}
\description{
function to link look-ups to the SAV file in dataframe form
}
