% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bias_CREST.R
\name{Load_C_Bias}
\alias{Load_C_Bias}
\title{Function to load bias corrections as from the tool that generates them into CREST
parameters are:
datapath = path w/filename to the csv file containing the corrections
dbDest = the designation of the destintion database be it "P" for OIOSP01 (prod) or "T" for OIOST01 (test)
or "D" for OISOD01 (dev)}
\usage{
Load_C_Bias(datapath, dbDest)
}
\description{
Function to load bias corrections as from the tool that generates them into CREST
parameters are:
datapath = path w/filename to the csv file containing the corrections
dbDest = the designation of the destintion database be it "P" for OIOSP01 (prod) or "T" for OIOST01 (test)
or "D" for OISOD01 (dev)
}
