% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates_CREST.R
\name{Load_C_Ests}
\alias{Load_C_Ests}
\title{function to wrap together all the steps needed to go from iREC analysis tool output to a
set of database tables
parameters are:
xls_path = file path to the source data
xls_file = the file name of the source data. file name needs to be month_year_ResultXXXX.xlsx
dbDest = the oracle database we are going to use to write the data to - P(prod) or T(Test)}
\usage{
Load_C_Ests(xls_pathIn, xls_fileName, dbDest, session)
}
\description{
function to wrap together all the steps needed to go from iREC analysis tool output to a
set of database tables
parameters are:
xls_path = file path to the source data
xls_file = the file name of the source data. file name needs to be month_year_ResultXXXX.xlsx
dbDest = the oracle database we are going to use to write the data to - P(prod) or T(Test)
}
