% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_Code.R
\name{load_CE_lus}
\alias{load_CE_lus}
\title{Function to load look up tables from CREST
parameters are:
df_tbl_nmes = list of table names to load}
\usage{
load_CE_lus(ccon, df_tbl_nmes)
}
\description{
Function to load look up tables from CREST
parameters are:
df_tbl_nmes = list of table names to load
}
