% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses_CREST.R
\name{load_CR_lus}
\alias{load_CR_lus}
\title{function to load look-ups from CREST to aid in formatting response data for loading into CREST}
\usage{
load_CR_lus()
}
\description{
function to load look-ups from CREST to aid in formatting response data for loading into CREST
}
