% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bias_KREST.R
\name{Load_K_Bias}
\alias{Load_K_Bias}
\title{wrapper function to process a set of bias correction factors in CSV format as provided by the
analysis tool}
\usage{
Load_K_Bias(datapath, dbDest)
}
\description{
wrapper function to process a set of bias correction factors in CSV format as provided by the
analysis tool
}
