% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates_KREST.R
\name{Load_K_Ests}
\alias{Load_K_Ests}
\title{wrapper function to send and promote an estimate packet to KREST}
\usage{
Load_K_Ests(xls_path, xls_file, source_id, session)
}
\description{
wrapper function to send and promote an estimate packet to KREST
}
