% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates_KREST.R
\name{loadCSVs}
\alias{loadCSVs}
\title{function to load a set of csv files into a list of dataframes
3 of the files will be "refreshed" from the CREST database if there is
a known connection preexisting
parameters are:
none
return a list of dataframes}
\usage{
loadCSVs()
}
\description{
function to load a set of csv files into a list of dataframes
3 of the files will be "refreshed" from the CREST database if there is
a known connection preexisting
parameters are:
none
return a list of dataframes
}
