% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_Main.R
\name{LoadNeighbours}
\alias{LoadNeighbours}
\title{Function to load the table of areas fished in a sigle day that are not flagged.
figuratively which areas "neighbours", but in practice, there may be gaps between
parameters are:
neighbours_df = dataframe loaded from csv of each area and a csv list of acceptable "neighbours"
areas_df = dataframe of iREC areas and their ref data names}
\usage{
LoadNeighbours(neighbours_df, areas_df)
}
\description{
Function to load the table of areas fished in a sigle day that are not flagged.
figuratively which areas "neighbours", but in practice, there may be gaps between
parameters are:
neighbours_df = dataframe loaded from csv of each area and a csv list of acceptable "neighbours"
areas_df = dataframe of iREC areas and their ref data names
}
