% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_Main.R
\name{LoadSav}
\alias{LoadSav}
\title{Function to take the raw SAV file eoncvert it to current standards if necessary and apply the excludes and adjusts
parameters are:
SavDat = the SAV file loaded using haven::read_sav
excludes - the csv file containing the licences to exclude
adjusts = the csv file containing the adjusted records}
\usage{
LoadSav(SavDat, excludes, adjusts)
}
\description{
Function to take the raw SAV file eoncvert it to current standards if necessary and apply the excludes and adjusts
parameters are:
SavDat = the SAV file loaded using haven::read_sav
excludes - the csv file containing the licences to exclude
adjusts = the csv file containing the adjusted records
}
