% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common_UniformSav.R
\name{loadSurveyAdjustments}
\alias{loadSurveyAdjustments}
\title{Load Survey Adjustments}
\usage{
loadSurveyAdjustments(adjusts)
}
\arguments{
\item{survey_adj_filename}{CSV file name of survey adjustments}
}
\value{
A tibble of updated survey data to replace original survey results
}
\description{
Loads and validates the survey adjustment file
}
