% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates_CREST.R
\name{loadXLSX_CREST}
\alias{loadXLSX_CREST}
\title{Function to read selected sheets from and XLSX file into 1 data frame for each
parameters are:
xls_path = the path to the file
xls_file = the xlsx file name
SHEET_NAMES = a list of named tabs in the XLSX file to load
returns a list of data frames named for the xlsx tab}
\usage{
loadXLSX_CREST(xls_path, xls_file, SHEET_NAMES)
}
\description{
Function to read selected sheets from and XLSX file into 1 data frame for each
parameters are:
xls_path = the path to the file
xls_file = the xlsx file name
SHEET_NAMES = a list of named tabs in the XLSX file to load
returns a list of data frames named for the xlsx tab
}
