% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates_KREST.R
\name{loadXLSX_KREST}
\alias{loadXLSX_KREST}
\title{function to load iREC estimates into dataframes
also to perform some standardization as the iREC analysis output has changed over time
this functions to make the source data equal to the current analysis output
this can use a file on the fil system or pull the file from the CREST database table
OTOLITH_V1.CREEL_IREC_SOURCE - this extra function is useful for transferring the
already published estimates in CREST to KREST without finding the source xlsx files
paramaters are:
xls_file = the filename with no path
xls_paths = a list of paths to search for the file named above with trailing slash. - Not sure why i did it this way :-P
inds = the components of the file name in the form Month_Year_... like "May_2026_Result_20260723_092514.xlsx"
this is to readily identify the time period the data represents.  The iREC analysis output does not identify the year or the month
it is only in the file name
SHEET_NAMES = a list of the sheet(tabs) names to pull for the analysis
source_id = the primary key of the record in the table OTOLITH_V1.CREEL_IREC_SOURCE to fetch the source excel file from}
\usage{
loadXLSX_KREST(xls_file, xls_paths, inds, SHEET_NAMES, source_id)
}
\description{
function to load iREC estimates into dataframes
also to perform some standardization as the iREC analysis output has changed over time
this functions to make the source data equal to the current analysis output
this can use a file on the fil system or pull the file from the CREST database table
OTOLITH_V1.CREEL_IREC_SOURCE - this extra function is useful for transferring the
already published estimates in CREST to KREST without finding the source xlsx files
paramaters are:
xls_file = the filename with no path
xls_paths = a list of paths to search for the file named above with trailing slash. - Not sure why i did it this way :-P
inds = the components of the file name in the form Month_Year_... like "May_2026_Result_20260723_092514.xlsx"
this is to readily identify the time period the data represents.  The iREC analysis output does not identify the year or the month
it is only in the file name
SHEET_NAMES = a list of the sheet(tabs) names to pull for the analysis
source_id = the primary key of the record in the table OTOLITH_V1.CREEL_IREC_SOURCE to fetch the source excel file from
}
