% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates_KREST.R
\name{norm_Data}
\alias{norm_Data}
\title{function to normalize a data table based on a list of column names
this isn't universal - it expects a column tmp_id and method at the least
it also does some double duty and forces there to be at least an estimate of 0 for species
matched in the vector "pattern" - there is a tendency to assume that no estimate means there wasn't a
survey and to go looking eslewere for an estimate.
parameters are:
items = list of column names to normalize.  This column codes in the disposition, mark status, size, etc
we also break all those potentials out into their own fields
dat_tbl = the data frame to work on
returns dataframe with minimal id's and item/value pair}
\usage{
norm_Data(items, data_tbl)
}
\description{
function to normalize a data table based on a list of column names
this isn't universal - it expects a column tmp_id and method at the least
it also does some double duty and forces there to be at least an estimate of 0 for species
matched in the vector "pattern" - there is a tendency to assume that no estimate means there wasn't a
survey and to go looking eslewere for an estimate.
parameters are:
items = list of column names to normalize.  This column codes in the disposition, mark status, size, etc
we also break all those potentials out into their own fields
dat_tbl = the data frame to work on
returns dataframe with minimal id's and item/value pair
}
