% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_Main.R
\name{processSav}
\alias{processSav}
\title{function to process the iREC sav file as delivered by th econtractor
parameters are:
SavDat = Name (and path?) to the .sav file
xlsFile = Name of the output XLSX file - THis might be a bit of legacy?
HABAG = the Daily Halibut Bag limit if not using FRIS directly
CNBAG = the Daily Chinook Bag limit if not using FRIS directly
COBAG = the Daily COHO Bag limit if not using FRIS directly
ENCBAG = the Daily Encounter Bag limit used for all three species
ANN_CN = the Annual limit for retained Chinook
ANN_LC = the Annual limit for retained Lingcod
ANN_HA = the Annual limit for retained Halibut
excludes = the dataframe holding the exclude data
adjusts = the dataframe holding the adjusts data
areaout_returns = Boolean flag to return all repeating area patters, or just the summer excludes
FRIS_USE = Boolean flag to pull bag limits data from FRIS or use defaults
FRIS_GRACE = integer to add to the FRIS daily limits before flagging the record}
\usage{
processSav(
  SavDat,
  xlsFile,
  HABAG,
  CNBAG,
  COBAG,
  ENCBAG,
  ANN_CN,
  ANN_LC,
  ANN_HA,
  excludes,
  adjusts,
  areaout_returns,
  FRIS_USE,
  FRIS_GRACE
)
}
\description{
function to process the iREC sav file as delivered by th econtractor
parameters are:
SavDat = Name (and path?) to the .sav file
xlsFile = Name of the output XLSX file - THis might be a bit of legacy?
HABAG = the Daily Halibut Bag limit if not using FRIS directly
CNBAG = the Daily Chinook Bag limit if not using FRIS directly
COBAG = the Daily COHO Bag limit if not using FRIS directly
ENCBAG = the Daily Encounter Bag limit used for all three species
ANN_CN = the Annual limit for retained Chinook
ANN_LC = the Annual limit for retained Lingcod
ANN_HA = the Annual limit for retained Halibut
excludes = the dataframe holding the exclude data
adjusts = the dataframe holding the adjusts data
areaout_returns = Boolean flag to return all repeating area patters, or just the summer excludes
FRIS_USE = Boolean flag to pull bag limits data from FRIS or use defaults
FRIS_GRACE = integer to add to the FRIS daily limits before flagging the record
}
