% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_Code.R
\name{PurgeDte}
\alias{PurgeDte}
\title{function to extract the occurance of a value from a dataframe column - used in figuring out what a date is
parameters are:
df = the dataframe holding the date in question
val = the value we are looking for
frst = boolean - to limit the check to the first occasion (maybe?)}
\usage{
PurgeDte(df, val, frst)
}
\description{
function to extract the occurance of a value from a dataframe column - used in figuring out what a date is
parameters are:
df = the dataframe holding the date in question
val = the value we are looking for
frst = boolean - to limit the check to the first occasion (maybe?)
}
