% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates_KREST.R
\name{purgeEstWoptions}
\alias{purgeEstWoptions}
\title{function to purge an estimate by its base name
parameters are:
estname = the base of the estimate name
krestroot = the KREST url}
\usage{
purgeEstWoptions(estname, krestroot)
}
\description{
function to purge an estimate by its base name
parameters are:
estname = the base of the estimate name
krestroot = the KREST url
}
