% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates_KREST.R
\name{read_all_sheets}
\alias{read_all_sheets}
\title{function to read in all the sheets in and excel file
parameters are:
xlsxFile = file with path to the excel file to read
returns a named list of all the sheets}
\usage{
read_all_sheets(xlsxFile, ...)
}
\description{
function to read in all the sheets in and excel file
parameters are:
xlsxFile = file with path to the excel file to read
returns a named list of all the sheets
}
