% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_Code.R
\name{read_file_bytes}
\alias{read_file_bytes}
\title{Function to read in a binary file
mostly to store files in an Oracle database
parameters are:
file_path = path + filename to the file to read}
\usage{
read_file_bytes(file_path)
}
\description{
Function to read in a binary file
mostly to store files in an Oracle database
parameters are:
file_path = path + filename to the file to read
}
