% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iRECui.R
\name{run_iRECUI}
\alias{run_iRECUI}
\title{Run iRec User Interface}
\usage{
run_iRECUI(port = 9997)
}
\arguments{
\item{port}{Port number to run shiny server on}
}
\description{
Start a shiny server with the user interface for the iRecUI package
}
