% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bias_KREST.R
\name{SendBiasiRECD}
\alias{SendBiasiRECD}
\title{function to delete a bias correction data packet in KREST based on GUID}
\usage{
SendBiasiRECD(krestendpoint, authToken, guid)
}
\description{
function to delete a bias correction data packet in KREST based on GUID
}
