% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bias_KREST.R
\name{SendBiasiRECPro}
\alias{SendBiasiRECPro}
\title{function to "promote" a bias correction data packet in KREST based on GUID
parameters are:
krestendpoint = the krest server endpoint we want to talk to
authToken = the authorization token used to authenticate
guid = the guid of the estimate we want to promote}
\usage{
SendBiasiRECPro(krestendpoint, authToken, guid)
}
\description{
function to "promote" a bias correction data packet in KREST based on GUID
parameters are:
krestendpoint = the krest server endpoint we want to talk to
authToken = the authorization token used to authenticate
guid = the guid of the estimate we want to promote
}
