% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bias_KREST.R
\name{SendBiasJSON}
\alias{SendBiasJSON}
\title{function wrapper to send and promote bias correction data in/to KREST}
\usage{
SendBiasJSON(Jsons, krestendpoint, login)
}
\description{
function wrapper to send and promote bias correction data in/to KREST
}
