% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates_KREST.R
\name{SendiRECP}
\alias{SendiRECP}
\title{function to send an estimate packet by its GUID
parameters are:
krestendpoint = the KREST url
authToken = and authorization token from KREST
bodyj = the JSON data packet to send}
\usage{
SendiRECP(krestendpoint, authToken, bodyj)
}
\description{
function to send an estimate packet by its GUID
parameters are:
krestendpoint = the KREST url
authToken = and authorization token from KREST
bodyj = the JSON data packet to send
}
