% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates_KREST.R
\name{SendiRECPro}
\alias{SendiRECPro}
\title{function to promote an estimate by its GUID
parameters are:
krestendpoint = the KREST url
authToken = and authorization token from KREST
GUID the GUID of the estimate to delete}
\usage{
SendiRECPro(krestendpoint, authToken, guid)
}
\description{
function to promote an estimate by its GUID
parameters are:
krestendpoint = the KREST url
authToken = and authorization token from KREST
GUID the GUID of the estimate to delete
}
