% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates_KREST.R
\name{SendJSON}
\alias{SendJSON}
\title{function to take the estimate data packages with the JSON string and send them to KREST
via a HTTPS web call
parameters are:
Jsons = the data frame of the estimates to send including a field containing the JSON string
krestendpoint = the url of the destination server
logincreds = the account credentials used to send the data to KREST}
\usage{
SendJSON(Jsons, krestendpoint, logincreds, krestroot)
}
\description{
function to take the estimate data packages with the JSON string and send them to KREST
via a HTTPS web call
parameters are:
Jsons = the data frame of the estimates to send including a field containing the JSON string
krestendpoint = the url of the destination server
logincreds = the account credentials used to send the data to KREST
}
