% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates_CREST.R
\name{SendToCREST}
\alias{SendToCREST}
\title{Function to take the fully formated estimate data and push it to the database (CREST)
we use the dbWriteTable function which is crazy faster than individual inserts, but all columns in the dataframe must match
the fields in teh database table exactly
parameters are:
rslts = list containing the 4 tables creel_irec_Est, creel_irec_value, creel_irec_var, creel_irec_var_value
Crestcon = oracle connection}
\usage{
SendToCREST(rslts, Crestcon)
}
\description{
Function to take the fully formated estimate data and push it to the database (CREST)
we use the dbWriteTable function which is crazy faster than individual inserts, but all columns in the dataframe must match
the fields in teh database table exactly
parameters are:
rslts = list containing the 4 tables creel_irec_Est, creel_irec_value, creel_irec_var, creel_irec_var_value
Crestcon = oracle connection
}
