% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates_KREST.R
\name{setFilter}
\alias{setFilter}
\title{function to set a view filter in the KREST UI
using the function FetchGrid we can get a dataframe of the available estimates that match the name
of interest.  We are pretending to be using the UI, but are just using the UIs web calls
used for determining if an estimate exists effectively with a wildcard
useful for when an estimate has been split i.e. GST-1, GST-2
there is no guarentee that a subsequent estimate will have the same number of splits
and not checking first could leave orphaned splits in the database
parameters are:
est_name = the base name of the estimate to search for
krestroot = the database we are interested in.}
\usage{
setFilter(est_name, krestroot)
}
\description{
function to set a view filter in the KREST UI
using the function FetchGrid we can get a dataframe of the available estimates that match the name
of interest.  We are pretending to be using the UI, but are just using the UIs web calls
used for determining if an estimate exists effectively with a wildcard
useful for when an estimate has been split i.e. GST-1, GST-2
there is no guarentee that a subsequent estimate will have the same number of splits
and not checking first could leave orphaned splits in the database
parameters are:
est_name = the base name of the estimate to search for
krestroot = the database we are interested in.
}
