% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_adjust_validation_v2.R
\name{setupOAV}
\alias{setupOAV}
\title{function to setup the code and variables for validating the adjustment data}
\usage{
setupOAV()
}
\description{
function to setup the code and variables for validating the adjustment data
}
