% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_Code.R
\name{setupOracleConn}
\alias{setupOracleConn}
\title{Function to build a connection to one of the oracle databases iREC references
parameter "S" indicates which server to attach to .
S = P means Production server for the OTOLITH_V1. schema (where CREST data (and iREC) are stored)
S = T  means Test server for the OTOLITH_V1. schema (where CREST data (and iREC) are stored)
S = L  means the ReadOnly LakeHose account to get data from FRIS and KREST
credentials are stored in a text file in the user's Documents folder in a subfolder named OracleCreds}
\usage{
setupOracleConn(S)
}
\description{
Function to build a connection to one of the oracle databases iREC references
parameter "S" indicates which server to attach to .
S = P means Production server for the OTOLITH_V1. schema (where CREST data (and iREC) are stored)
S = T  means Test server for the OTOLITH_V1. schema (where CREST data (and iREC) are stored)
S = L  means the ReadOnly LakeHose account to get data from FRIS and KREST
credentials are stored in a text file in the user's Documents folder in a subfolder named OracleCreds
}
