% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_adjust_validation_v2.R
\name{standardize_crabprawn}
\alias{standardize_crabprawn}
\title{---- Helper: standardize old crab/prawn columns ----
Function to standardize the crab/prawn fields over time}
\usage{
standardize_crabprawn(df)
}
\description{
---- Helper: standardize old crab/prawn columns ----
Function to standardize the crab/prawn fields over time
}
