% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates_CREST.R
\name{StoreSource}
\alias{StoreSource}
\title{Function to zip up the estiamte xlsx file and store it in the database
returning the source_id that was used
prameters are:
dbCon = oracle connection with write access to IREC tables
datafile = name of the XLSX file to store
datpath = path to the XLSX file to store
yr_name = the name of the fiscal year the estimates are part of
mnth_str = the name of the month the estimates are for}
\usage{
StoreSource(dbCon, datafile, datapath, yr_name, mnth_str)
}
\description{
Function to zip up the estiamte xlsx file and store it in the database
returning the source_id that was used
prameters are:
dbCon = oracle connection with write access to IREC tables
datafile = name of the XLSX file to store
datpath = path to the XLSX file to store
yr_name = the name of the fiscal year the estimates are part of
mnth_str = the name of the month the estimates are for
}
