% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common_UniformSav.R
\name{UnspecifiedText}
\alias{UnspecifiedText}
\title{this code is almost entirely Nick's  Just copied here to allow for converting EKOS to PRA data fromat
Convert SPSS labels to text}
\usage{
UnspecifiedText
}
\description{
this code is almost entirely Nick's  Just copied here to allow for converting EKOS to PRA data fromat
Convert SPSS labels to text
}
