% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_adjust_validation_v2.R
\name{validate_adjustments}
\alias{validate_adjustments}
\title{---- adjustment Validation function ----}
\usage{
validate_adjustments(adjusts, excludes = NULL, excel_file = NULL)
}
\description{
---- adjustment Validation function ----
}
