% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses_CREST.R
\name{write_C_ExcludesIn}
\alias{write_C_ExcludesIn}
\title{function to load excluded data into the CREST database
parameters are:
datapath = the location of the source exclude file including file name
datafile = the source name of the file - shiny renames the files 0.csv when transferring them, so we need to
know the real name
yr_name = the licence year the file is to be applied to
df_read = the dataframe as loaded from the source file}
\usage{
write_C_ExcludesIn(datapath, datafile, yr_name, df_read)
}
\description{
function to load excluded data into the CREST database
parameters are:
datapath = the location of the source exclude file including file name
datafile = the source name of the file - shiny renames the files 0.csv when transferring them, so we need to
know the real name
yr_name = the licence year the file is to be applied to
df_read = the dataframe as loaded from the source file
}
