% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses_CREST.R
\name{Write_C_ResponsesIn}
\alias{Write_C_ResponsesIn}
\title{function wrapper to load responses and adjusts into CREST
datapath = the location of the source exclude file including file name
datafile = the source name of the file - shiny renames the files 0.csv when transferring them, so we need to
know the real name
yr_name = the licence year the file is to be applied to
yr = the year the data is from
mnth_str = the string name for the month the data is from
df_read = the dataframe as loaded from the source file
datatype = either "Response" or "Adjust" for the type of data to be inserted}
\usage{
Write_C_ResponsesIn(
  datapath,
  datafile,
  yr_name,
  yr,
  mnth_str,
  df_read,
  datatype
)
}
\description{
function wrapper to load responses and adjusts into CREST
datapath = the location of the source exclude file including file name
datafile = the source name of the file - shiny renames the files 0.csv when transferring them, so we need to
know the real name
yr_name = the licence year the file is to be applied to
yr = the year the data is from
mnth_str = the string name for the month the data is from
df_read = the dataframe as loaded from the source file
datatype = either "Response" or "Adjust" for the type of data to be inserted
}
