% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses_KREST.R
\name{Write_K_Responses}
\alias{Write_K_Responses}
\title{wrapper function to load response data into KREST}
\usage{
Write_K_Responses(sourcepath, destpath, datapath, datafile, yr_name)
}
\description{
wrapper function to load response data into KREST
}
