% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses_KREST.R
\name{Write_K_ResponsesIn}
\alias{Write_K_ResponsesIn}
\title{wrapper function to load response data into KREST}
\usage{
Write_K_ResponsesIn(sourcepath, destpath, datapath, datafile, yr_name, df_read)
}
\description{
wrapper function to load response data into KREST
}
