% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bias_KREST.R
\name{buildBiasJSONS}
\alias{buildBiasJSONS}
\title{function to create a JSON dat packet that KREST can injest}
\usage{
buildBiasJSONS(rslt, lic_str, csv_file)
}
\description{
function to create a JSON dat packet that KREST can injest
}
