% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KREST_common.R
\name{GetLookupCodes}
\alias{GetLookupCodes}
\title{function to fetch look-up tables from KREST
parameters are:
BearerToken = the access token from getAuthToken
resource = the lookup table in question
dateTime = optional point in time selection
showAssignementHistory = not realluy sure.  Maybe is supposed to return when a value was added?}
\usage{
GetLookupCodes(
  BearerToken,
  resource,
  dateTime = NULL,
  showAssignmentHistory = FALSE
)
}
\description{
function to fetch look-up tables from KREST
parameters are:
BearerToken = the access token from getAuthToken
resource = the lookup table in question
dateTime = optional point in time selection
showAssignementHistory = not realluy sure.  Maybe is supposed to return when a value was added?
}
