% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_Code.R
\name{linkTables}
\alias{linkTables}
\title{function to link look up tables to response data}
\usage{
linkTables(survey_data, fullshellfish)
}
\description{
function to link look up tables to response data
}
