% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_Code.R
\name{load_resp}
\alias{load_resp}
\title{Function to load response files for either CREST or KREST format
based on the destination}
\usage{
load_resp(inFile, Dest)
}
\description{
Function to load response files for either CREST or KREST format
based on the destination
}
