% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates_KREST.R
\name{LoadxlsxFromCREST}
\alias{LoadxlsxFromCREST}
\title{function to load an excel estimate as provided by the anaysis package}
\usage{
LoadxlsxFromCREST(source_id)
}
\description{
function to load an excel estimate as provided by the anaysis package
}
