% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bias_KREST.R
\name{SendBiasiRECP}
\alias{SendBiasiRECP}
\title{function to "post" a bias correction data packet to KREST}
\usage{
SendBiasiRECP(krestendpoint, authToken, bodyj, updateEstimate)
}
\description{
function to "post" a bias correction data packet to KREST
}
